ci: make kas checkout correct ref
When kas goes to get the configuration files it uses to build,
it defaults to checking out main. Normally, this is a good thing,
but when we're trying to run CI on a PR, this is problematic (we
need to test our change). So, generate a .yml on the fly with the
correct ref and use kas to concatenate those files.

Signed-off-by: Dillon Wells <dillon.wells@cesiumastro.com>